refactor: Update styles.js file names to follow new convention
