--- california-dev/src/components/ui/HamburgerMenu/HamburgerMenu.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-import { HamburgerContainer, Line } from "./styles";
-export const HamburgerMenu = () => {
-  return (
-    <HamburgerContainer>
-      <Line width="3.2rem" />
-      <Line width="2.0rem" />
-      <Line width="1.6rem" />
-    </HamburgerContainer>
-  );
-};
