--- california-dev/src/components/ui/HamburgerMenu/index.js
@@ -1,1 +1,0 @@
-export * from "./HamburgerMenu";
