+++ california-dev/src/components/ui/MenuButton/MenuButton.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { ButtonWrapper, Line } from "./MenuButtonStyles";
+export const MenuButton = () => {
+  return (
+    <ButtonWrapper>
+      <Line width="3.2rem" />
+      <Line width="2.0rem" />
+      <Line width="1.6rem" />
+    </ButtonWrapper>
+  );
+};
